hotfix/CA-030 unificar estilos de componentes
