+++ decide/booth/decide-react/src/components/Barra.js
@@ -2,7 +2,7 @@ import React, {Component} from 'react';
 import { StatusBar, Text, View } from 'react-native';
 import { postData } from '../utils';
 import config from '../config.json';
-import { StyleSheet} from "react-native";
+import { styles } from "../styles";
 
 export default class Barra extends Component{
 
@@ -34,29 +34,3 @@ export default class Barra extends Component{
         );
     }
 }
-
-
-const styles = StyleSheet.create ({
-
-    barraStyle: {
-        width: '100%',
-        backgroundColor: '#002080',
-        justifyContent: 'space-between',
-        paddingHorizontal: 20,
-        paddingTop: 15,
-        paddingBottom: 15,
-        display: 'flex',
-        flexDirection: 'row',
-        alignItems: 'center',
-    },
-    titleStyle: {
-        color: 'white',
-        fontSize: 18,
-    },
-    textStyle: {
-        color: 'white',
-        fontSize: 14,
-    },
-
-});
-
