+++ decide/booth/decide-react/src/components/Voting.js
@@ -6,6 +6,7 @@ import { postData } from '../utils';
 import config from '../config.json';
 import { StyleSheet} from "react-native";
 import RadioForm from 'react-native-simple-radio-button';
+import { styles } from "../styles";
 
 export default class Voting extends Component {
 
@@ -138,138 +139,3 @@ export default class Voting extends Component {
       </View>;      
     }
 }
-
-const styles = StyleSheet.create ({
-  html: {
-    margin: 0,
-    padding: 0,
-  },
-  body: {
-      margin: 0,
-      padding: 0,
-      fontFamily: '"Helvetica Neue",Helvetica,Arial,sans-serif',
-      fontSize: 18,
-      fontWeight: 'normal',
-      lineHeight: 24,
-      display: 'flex',
-      alignItems: 'center',
-      alignContent: 'center',
-      backgroundColor: '#fff',
-  },
-  container: {
-      width: '100%',
-      maxWidth: 960,
-      justifyContent: 'center',
-      alignItems: 'center',
-  },
-  content: {
-      width: '100%',
-      borderRadius: 10,
-      padding: 25,
-  },
-    containerStyle: {
-        justifyContent: 'center',
-        alignItems: 'center'
-    },
-    contentStyle: {
-        width: 960,
-        backgroundColor: '#fff',
-        borderTopLeftRadius: 10,
-        borderTopRightRadius: 10,
-        borderBottomRightRadius: 10,
-        borderBottomLeftRadius: 10,
-        overflow: 'hidden',
-        display: 'flex',
-        flexWrap: 'wrap',
-        justifyContent: 'space-between',
-        paddingTop: 50,
-        paddingRight: 50,
-        paddingBottom: 50,
-        paddingLeft: 50
-    },
-    textStyle:{
-        fontSize: 26,
-        lineHeight: 24,
-        paddingTop:10, 
-        paddingBottom:10,
-        justifyContent: 'center',
-        alignSelf: 'center'
-    },
-    radioStyle:{
-        justifyContent: 'center',
-        alignSelf: 'center'  
-    },
-    row: {
-    },
-    clearfix: {
-        "marginBottom": 24,
-    },
-    button1Style: {
-        width: "100%",
-        display: "flex",
-        flexWrap: "wrap",
-        justifyContent: "center",
-        fontSize: 18,
-        lineHeight: 1.5,
-        color: "#fff",
-        textTransform: "uppercase",
-        height: 50,
-        borderTopLeftRadius: 25,
-        borderTopRightRadius: 25,
-        borderBottomRightRadius: 25,
-        borderBottomLeftRadius: 25,
-        backgroundColor: "#0064cd",
-        paddingTop: 0,
-        paddingRight: 25,
-        paddingBottom: 0,
-        paddingLeft: 25,
-        textShadowOffset: {
-          width: 0,
-          height: -1
-        },
-        textShadowRadius: 0,
-        textShadowColor: "rgba(0, 0, 0, 0.25)",
-        borderTopColor: "#0064cd",
-        borderRightColor: "#0064cd",
-        borderBottomColor: "#0064cd",
-        borderLeftColor: "#0064cd"
-    },
-    button2Style: {
-        width: "100%",
-        display: "flex",
-        flexWrap: "wrap",
-        justifyContent: "center",
-        fontSize: 18,
-        lineHeight: 1.5,
-        color: "#fff",
-        textTransform: "uppercase",
-        height: 50,
-        borderTopLeftRadius: 25,
-        borderTopRightRadius: 25,
-        borderBottomRightRadius: 25,
-        borderBottomLeftRadius: 25,
-        backgroundColor: "#000000",
-        paddingTop: 0,
-        paddingRight: 25,
-        paddingBottom: 0,
-        paddingLeft: 25,
-        textShadowOffset: {
-          width: 0,
-          height: -1
-        },
-        textShadowRadius: 0,
-        textShadowColor: "rgba(0, 0, 0, 0.25)",
-        borderTopColor: "#000000",
-        borderRightColor: "#000000",
-        borderBottomColor: "#000000",
-        borderLeftColor: "#000000"
-    },
-    votingStyle:{
-      fontSize: 26,
-      lineHeight: 24,
-      paddingTop:10, 
-      paddingBottom:10,
-      justifyContent: 'center',
-      alignSelf: 'center'
-  },
-});
